+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Enable parallel tests on Windows
 - Add `assert_not_called`
 - Improve `find_total_tests` performance
+- Improve `clock` performance
 
 ## [0.19.1](https://github.com/TypedDevs/bashunit/compare/0.19.0...0.19.1) - 2025-05-23